+++ docs/changelog.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## Current
+
+## V1.2.0
+
+* feat(mail): Adds new config to send custom `from` and specify a template to be
+used as an email body when resetting a password, or confirming a new account.
+The URL for mailer has also been made configurable. (ASG-2630)
+
 ## v1.1.2
 
 * fix: Update authaus version. (ASG-2622)